feat: Support markdown and PDF files
