+++ src/components/KnowledgeStoreUploader.ts
@@ -6,8 +6,6 @@
 // // State
 // import { useKnowledgeStore } from 'src/stores/knowledge-store';
 //
-// // Get PDF.js from the window object
-// const pdfjsLib = window.pdfjsLib;
 //
 // export type AttachmentAddedEvent = {
 //   title: string;
@@ -103,64 +101,7 @@
 //       fileStatus.value = {};
 //     }
 //
-//     async function processFile(file: File) {
-//       const title = file.name;
-//       let extractedText = '';
-//       let type = file.type;
 //
-//       try {
-//         switch (file.type) {
-//           case 'application/pdf':
-//             extractedText = await extractTextFromPdfFile(file);
-//             break;
-//           case 'text/plain':
-//             extractedText = await new Promise((resolve, reject) => {
-//               const reader = new FileReader();
-//               reader.onload = (event) => resolve(event.target!.result as string);
-//               reader.onerror = (error) => reject(error);
-//               reader.readAsText(file);
-//             });
-//             break;
-//           case 'text/markdown':
-//             extractedText = await new Promise((resolve, reject) => {
-//               const reader = new FileReader();
-//               reader.onload = (event) => resolve(event.target!.result as string);
-//               reader.onerror = (error) => reject(error);
-//               reader.readAsText(file);
-//             });
-//             break;
-//           default:
-//             throw new Error(`Unsupported file type: ${file.type}`);
-//         }
-//       } catch (error) {
-//         console.error('Error processing file:', error);
-//         throw error;
-//       }
-//
-//       return { title, text: extractedText, type };
-//     }
-//
-//     async function extractTextFromPdfFile(file: File) {
-//       const pdfUrl = URL.createObjectURL(file);
-//
-//       let pdf;
-//       try {
-//         pdf = await pdfjsLib.getDocument(pdfUrl).promise;
-//       } catch (error) {
-//         console.error(`components::KnowledgeStoreUploader::extractTextFromPdfFile - error: ${error}`);
-//         throw new Error('Failed to extract text from PDF');
-//       }
-//       const maxPages = pdf.numPages;
-//       let textContent = [];
-//
-//       for (let i = 1; i <= maxPages; i++) {
-//         const page = await pdf.getPage(i);
-//         const content = await page.getTextContent();
-//         const pageTextContent = content.items.map((item) => item.str).join(' ');
-//         textContent.push(pageTextContent);
-//       }
-//       return textContent.join('');
-//     }
 //
 //     return {
 //       isUploading,
